+++ examples/rapier2d_colliders.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::needless_pass_by_value)]
 use bevy::{asset::LoadState, color::palettes::css, prelude::*};
-use bevy_collider_gen::{AbstractCollider, AbstractCollidersBuilder};
+use bevy_collider_gen::prelude::*;
 use bevy_prototype_lyon::{prelude::*, shapes};
 use bevy_rapier2d::prelude::*;
 use edges::Edges;
